requirements.txt
google-genai>=1.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
